Trying to copy relevant parts of old commit
