+++ app/src/config/constants.ts
@@ -12,3 +12,5 @@ export const ADOBE_EMBED_URL =
   appConfig.adobeEmbedUrl[appConfig.environment as ENV_KEY];
 
 export const TRUNCATED_LENGTH = 80;
+
+export const CARDS_PER_PAGE = 48;
